Unit testing support added.
RSpec support has been added for unit testing the Specify framework. This
ties into code coverage and reporting tools. There is a potential issue
if I have an acceptance test example that runs as part of the code
repository. The issue is that Specify will be hooking into RSpec to run
the acceptance tests and that may conflict with the standard unit test
execution performed by RSpec.