Further updates to build and reporting.
A clean task was added to remove all reports generated as part of
the build and test process.

A fix was made to the filter for RSpec. This was necessary to allow
accurate reporting of coverage information.

The code climate key was updated with what is apparently crucial
information to allow coverage metrics to be generated.